feat: 장소/충전소 분리 및 상세 페이지 ui 수정
