+++ app/src/main/java/com/jeju/evtravel/ui/detail/ChargerDetailScreen.kt
@@ -1,13 +1,38 @@
 package com.jeju.evtravel.ui.detail
 
-import androidx.compose.foundation.layout.*
-import androidx.compose.material3.*
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.statusBarsPadding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Divider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
+import com.jeju.evtravel.R
 import com.jeju.evtravel.domain.model.Place
 import com.jeju.evtravel.ui.theme.Variables
 
@@ -21,9 +46,38 @@ data class ChargerRowUi(
 fun isCharging(status: String): Boolean =
     status == "3" || status.equals("CHARGING", ignoreCase = true)
 
+private fun prettyType(code: String?): String {
+    // null, 공백, "0" 방어
+    val c = code?.trim()?.padStart(2, '0') ?: return ""
+    return when (c) {
+        "01" -> "DC차데모"
+        "02" -> "AC완속"
+        "03" -> "DC차데모+AC3상"
+        "04" -> "DC콤보"
+        "05" -> "DC차데모+DC콤보"
+        "06" -> "DC차데모+AC3상+DC콤보"
+        "07" -> "AC3상"
+        "08" -> "DC콤보(완속)"
+        "09" -> "NACS"
+        "10" -> "DC콤보+NACS"
+        else -> ""
+    }
+}
+
+private fun kwBucket(output: String?): String? {
+    val kw = output?.toDoubleOrNull()?.toInt() ?: return null
+    return when {
+        kw >= 200 -> "${kw} kW"          // 초급속 등
+        kw >= 50  -> "${kw} kW"          // 급속
+        kw > 0    -> "${kw} kW"          // 완속
+        else      -> null
+    }
+}
+
 @Composable
 fun ChargerDetailScreen(
     place: Place,                       // place.chargerList 사용
+    onBack: () -> Unit,
     onNavigateClick: () -> Unit
 ) {
     val chargers = place.chargerList ?: emptyList()
@@ -31,81 +85,128 @@ fun ChargerDetailScreen(
     // 샘플 변환: 출력(W)과 커넥터를 합쳐 라벨링하고 가격/가용수 계산
     val rows: List<ChargerRowUi> = chargers
         .groupBy { ch ->
-            // 출력 버킷 + 커넥터명(있으면)로 묶기
-            val kw = ch.output?.toDoubleOrNull()?.toInt() ?: 0
-            val bucket = when {
-                kw >= 100 -> "${kw} kWDC콤보"
-                kw in 50..99 -> "${kw} kWDC콤보"
-                else -> "${kw} kWAC"
-            }
-            bucket
+            val typeLabel = prettyType(ch.chargerType)              // <- chargerType 기반
+            val kwLabel = kwBucket(ch.output)                       // <- (옵션) 출력 병기
+            // 라벨 규칙: "타입 + (있으면) 공백 + kW"
+            if (kwLabel != null) "$kwLabel$typeLabel" else typeLabel
         }
-        .map { (bucket, list) ->
+        .map { (label, list) ->
             val total = list.size
             val avail = list.count { !isCharging(it.status) }
-//            val price = list.firstOrNull()?.price // 프로젝트 필드명에 맞게 조정
             ChargerRowUi(
-                label = bucket,
-//                priceText = if (price != null) "${price}원/kWh" else "-원/kWh",
+                label = label,
                 available = avail,
                 total = total
             )
         }
+        // 정렬: 개수 많은 순(원하면 다른 기준 추가 가능)
         .sortedByDescending { it.total }
 
     Scaffold(
         floatingActionButton = { ActionFAB(onClick = onNavigateClick) }
     ) { inner ->
-        Column(
+        LazyColumn(
             modifier = Modifier
-                .fillMaxSize()
                 .padding(inner)
+                .fillMaxSize()
+                .background(Color.White),
+            contentPadding = PaddingValues(bottom = 96.dp)
         ) {
-            HeaderImage(imageUrl = " ")
-            Spacer(Modifier.height(12.dp))
-
-            Text(
-                text = place.name,
-                style = DetailTitleTextStyle,
-                modifier = Modifier.padding(horizontal = 16.dp)
-            )
-            Spacer(Modifier.height(10.dp))
-
-            InfoCard(
-                hours = chargers.firstOrNull()?.usageTime ?: "24시간 이용가능",
-                address = place.address,
-                phone = place.phone
-            )
-
-            Spacer(Modifier.height(18.dp))
-
-            SectionTitle(
-                title = "충전기 정보",
-                modifier = Modifier.padding(horizontal = 16.dp)
-            )
-            Spacer(Modifier.height(10.dp))
-
-            // 행 카드들
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 16.dp),
-                verticalArrangement = Arrangement.spacedBy(10.dp)
-            ) {
-                rows.forEach { r ->
-                    ChargerRow(r)
-                }
-
-                if (rows.isEmpty()) {
-                    Text(
-                        "등록된 충전기 정보가 없습니다.",
-                        style = DetailSubtitleTextStyle,
-                        modifier = Modifier.padding(vertical = 8.dp)
-                    )
+            // 헤더(이미지/뒤로가기)
+            item {
+                Box(Modifier.fillMaxWidth()) {
+                    HeaderImage(imageUrl = " ")
+                    IconButton(
+                        onClick = onBack,
+                        modifier = Modifier
+                            .statusBarsPadding()
+                            .padding(3.dp)
+                            .align(Alignment.TopStart)
+                    ) {
+                        Icon(
+                            painter = painterResource(id = R.drawable.ic_back),
+                            contentDescription = "뒤로가기"
+                        )
+                    }
                 }
             }
+            // 상단 둥근 섹션
+            item {
+                Surface(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .offset(y = (-20).dp),
+                    shape = RoundedCornerShape(topStart = 20.dp, topEnd = 20.dp),
+                    color = Color.White
+                ) {
+                    Column(Modifier.fillMaxWidth()) {
+                        // ─ 기본 정보 ─
+                        Column(Modifier.padding(16.dp)) {
+                            Text(text = place.name, style = DetailTitleTextStyle)
+                            Spacer(Modifier.height(20.dp))
 
-            Spacer(Modifier.height(40.dp))
+                            val hours = chargers.firstOrNull()?.usageTime.orEmpty()
+                            val location = chargers.firstOrNull()?.location
+                                ?.takeIf { !it.isNullOrBlank() && it.lowercase() != "null" }
+                                .orEmpty()
+                            val addr = place.roadAddress?.takeIf { it.isNotBlank() } ?: place.address
+                            val fullAddress =
+                                if (!addr.isNullOrBlank())
+                                    if (location.isNotEmpty()) "${addr}, $location" else addr!!
+                                else
+                                    location
+
+                            if (hours.isNotBlank()) {
+                                InfoRow(iconRes = R.drawable.ic_clock, text = hours)
+                                Spacer(Modifier.height(12.dp))
+                            }
+                            if (fullAddress.isNotBlank()) {
+                                InfoRow(iconRes = R.drawable.ic_location, text = fullAddress)
+                                Spacer(Modifier.height(12.dp))
+                            }
+                            if (!place.phone.isNullOrBlank()) {
+                                InfoRow(iconRes = R.drawable.ic_phone, text = place.phone)
+                                Spacer(Modifier.height(18.dp))
+                            }
+                        }
+
+                        Divider(
+                            modifier = Modifier
+                                .fillMaxWidth()
+                                .height(7.dp),
+                            color = Variables.Grayscale50,
+                            thickness = 7.dp
+                        )
+                        Spacer(Modifier.height(18.dp))
+
+                        SectionTitle(
+                            title = "충전기 정보",
+                            modifier = Modifier.padding(horizontal = 16.dp)
+                        )
+                        Spacer(Modifier.height(12.dp))
+
+                        // ─ 행들 (Surface 안에서 렌더링: 한 화면처럼 보이게) ─
+                        Column(
+                            modifier = Modifier
+                                .fillMaxWidth()
+                                .padding(horizontal = 16.dp),
+                            verticalArrangement = Arrangement.spacedBy(10.dp)
+                        ) {
+                            rows.forEach { r -> ChargerRow(r) }
+
+                            if (rows.isEmpty()) {
+                                Text(
+                                    "등록된 충전기 정보가 없습니다.",
+                                    style = DetailOutputTextStyle,
+                                    modifier = Modifier.padding(vertical = 8.dp)
+                                )
+                            }
+                        }
+
+                        Spacer(Modifier.height(40.dp)) // 하단 여유
+                    }
+                }
+            }
         }
     }
 }
@@ -127,14 +228,14 @@ private fun ChargerRow(row: ChargerRowUi) {
                 Text(
 //                    text = row.label + " | " + row.priceText,
                     text = row.label + " | ",
-                    style = DetailLabelTextStyle.copy(fontWeight = FontWeight.SemiBold)
+                    style = DetailOutputTextStyle
                 )
             }
             val canUse = row.available > 0
             val rightText = if (canUse) "충전가능 ${row.available}/${row.total}"
             else "충전불가 ${row.available}/${row.total}"
             val rightColor = if (canUse) Variables.Blue700 else Color(0xFFD92B2B)
-            Text(text = rightText, color = rightColor, style = DetailLabelTextStyle)
+            Text(text = rightText, color = rightColor, style = DetailStatTextStyle)
         }
     }
 }
