+++ app/src/main/java/com/jeju/evtravel/ui/theme/Variables.kt
@@ -4,6 +4,9 @@ import androidx.compose.ui.graphics.Color
 
 object Variables {
     val Blue700: Color = Color(0xFF0173FF)
+
+    val Grayscale50: Color = Color(0xFFF5F5F5)
     val Grayscale300: Color = Color(0xFFC4C4C4)
+    val Grayscale400: Color = Color(0xFF9D9D9D)
     val Grayscale600: Color = Color(0xFF555555)
 }
