+++ app/src/main/res/drawable/ic_ai.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="15dp"
+    android:height="15dp"
+    android:viewportWidth="15"
+    android:viewportHeight="15">
+  <path
+      android:pathData="M5.315,2.336C4.916,2.22 4.62,2.252 4.404,2.368C4.189,2.485 4.001,2.716 3.879,3.114C3.757,3.515 3.716,4.05 3.775,4.696C3.785,4.801 3.797,4.908 3.812,5.017C4.335,4.884 4.897,4.779 5.488,4.706C5.897,4.132 6.327,3.614 6.763,3.166C6.741,3.148 6.719,3.13 6.697,3.113C6.189,2.711 5.718,2.453 5.315,2.336ZM7.5,2.477C7.439,2.426 7.379,2.377 7.318,2.329C6.746,1.876 6.16,1.54 5.594,1.376C5.025,1.211 4.441,1.211 3.928,1.489C3.415,1.767 3.095,2.256 2.923,2.823C2.751,3.387 2.713,4.061 2.779,4.788C2.795,4.958 2.816,5.132 2.844,5.31C2.674,5.371 2.51,5.435 2.353,5.503C1.683,5.79 1.108,6.145 0.694,6.565C0.278,6.987 0,7.5 0,8.084C0,8.667 0.278,9.181 0.694,9.603C1.108,10.022 1.683,10.377 2.353,10.665C2.51,10.732 2.674,10.796 2.844,10.857C2.816,11.035 2.795,11.21 2.779,11.38C2.713,12.106 2.751,12.781 2.923,13.344C3.095,13.911 3.415,14.401 3.928,14.678C4.441,14.956 5.025,14.957 5.594,14.791C6.16,14.627 6.746,14.291 7.318,13.838C7.379,13.79 7.439,13.741 7.5,13.69C7.561,13.741 7.621,13.79 7.682,13.838C8.254,14.291 8.84,14.627 9.406,14.791C9.975,14.957 10.559,14.956 11.072,14.678C11.585,14.401 11.905,13.911 12.077,13.344C12.249,12.781 12.287,12.106 12.221,11.38C12.205,11.21 12.184,11.035 12.156,10.857C12.326,10.796 12.49,10.732 12.647,10.665C13.317,10.377 13.892,10.022 14.306,9.603C14.722,9.181 15,8.667 15,8.084C15,7.5 14.722,6.987 14.306,6.565C13.892,6.145 13.317,5.79 12.647,5.503C12.49,5.435 12.326,5.371 12.156,5.31C12.183,5.132 12.205,4.958 12.221,4.788C12.287,4.061 12.249,3.387 12.077,2.823C11.905,2.256 11.585,1.767 11.072,1.489C10.559,1.211 9.975,1.211 9.406,1.376C8.84,1.54 8.254,1.876 7.682,2.329C7.621,2.377 7.561,2.426 7.5,2.477ZM7.5,3.842C7.275,4.072 7.05,4.324 6.828,4.597C7.049,4.588 7.274,4.584 7.5,4.584C7.726,4.584 7.951,4.588 8.172,4.597C7.95,4.324 7.725,4.072 7.5,3.842ZM9.512,4.705C9.102,4.132 8.673,3.614 8.237,3.166C8.259,3.148 8.281,3.13 8.302,3.113C8.811,2.711 9.282,2.453 9.685,2.336C10.084,2.22 10.38,2.252 10.596,2.368C10.811,2.485 10.999,2.716 11.12,3.114C11.242,3.515 11.284,4.05 11.225,4.696C11.215,4.801 11.203,4.908 11.188,5.017C10.665,4.884 10.103,4.779 9.512,4.705ZM8.949,5.65C8.484,5.607 7.999,5.584 7.5,5.584C7.001,5.584 6.516,5.607 6.051,5.65C5.791,6.039 5.539,6.454 5.302,6.893C5.085,7.293 4.891,7.691 4.72,8.084C4.891,8.476 5.085,8.875 5.302,9.274C5.539,9.713 5.791,10.129 6.051,10.517C6.516,10.561 7.001,10.584 7.5,10.584C7.999,10.584 8.484,10.561 8.949,10.517C9.209,10.129 9.46,9.713 9.698,9.274C9.915,8.875 10.109,8.476 10.28,8.084C10.109,7.691 9.915,7.293 9.698,6.893C9.46,6.454 9.209,6.039 8.949,5.65ZM10.773,6.788C10.71,6.664 10.644,6.54 10.578,6.417C10.47,6.218 10.359,6.023 10.246,5.832C10.506,5.881 10.757,5.938 10.998,6C10.935,6.256 10.86,6.519 10.773,6.788ZM10.773,9.379C10.71,9.503 10.644,9.627 10.577,9.751C10.47,9.95 10.359,10.145 10.246,10.335C10.506,10.286 10.757,10.229 10.998,10.167C10.935,9.911 10.86,9.648 10.773,9.379ZM11.953,9.866C11.808,9.292 11.611,8.693 11.365,8.084C11.611,7.474 11.808,6.875 11.953,6.301C12.056,6.34 12.156,6.381 12.253,6.422C12.849,6.677 13.3,6.969 13.594,7.267C13.886,7.563 14,7.839 14,8.084C14,8.329 13.886,8.604 13.594,8.9C13.3,9.198 12.849,9.49 12.253,9.745C12.156,9.787 12.056,9.827 11.953,9.866ZM11.188,11.151C10.665,11.284 10.103,11.389 9.512,11.462C9.102,12.036 8.673,12.553 8.237,13.002C8.259,13.019 8.281,13.037 8.303,13.054C8.811,13.457 9.282,13.714 9.685,13.831C10.084,13.947 10.38,13.916 10.596,13.799C10.811,13.682 10.999,13.451 11.12,13.053C11.243,12.652 11.284,12.117 11.225,11.471C11.215,11.367 11.203,11.26 11.188,11.151ZM7.5,12.325C7.725,12.095 7.95,11.843 8.172,11.57C7.951,11.579 7.726,11.584 7.5,11.584C7.274,11.584 7.049,11.579 6.828,11.57C7.05,11.843 7.275,12.095 7.5,12.325ZM4.754,10.335C4.641,10.145 4.53,9.95 4.422,9.751C4.355,9.627 4.29,9.503 4.227,9.379C4.14,9.648 4.065,9.911 4.002,10.167C4.243,10.229 4.494,10.286 4.754,10.335ZM3.812,11.151C4.335,11.283 4.897,11.389 5.488,11.462C5.897,12.036 6.327,12.553 6.763,13.002C6.741,13.019 6.719,13.037 6.697,13.054C6.189,13.457 5.718,13.714 5.315,13.831C4.916,13.947 4.62,13.916 4.404,13.799C4.188,13.682 4.001,13.451 3.879,13.053C3.757,12.652 3.716,12.117 3.775,11.471C3.785,11.367 3.797,11.26 3.812,11.151ZM3.047,9.866C3.192,9.292 3.388,8.693 3.635,8.084C3.388,7.474 3.192,6.875 3.047,6.301C2.944,6.34 2.844,6.381 2.747,6.422C2.151,6.677 1.7,6.969 1.406,7.267C1.114,7.563 1,7.839 1,8.084C1,8.329 1.114,8.604 1.406,8.9C1.7,9.198 2.151,9.49 2.747,9.745C2.844,9.787 2.944,9.827 3.047,9.866ZM4.002,6C4.065,6.256 4.14,6.519 4.227,6.788C4.29,6.664 4.355,6.54 4.422,6.417C4.53,6.218 4.641,6.023 4.754,5.832C4.494,5.882 4.243,5.938 4.002,6ZM6,8.084C6,7.255 6.672,6.584 7.5,6.584C8.328,6.584 9,7.255 9,8.084C9,8.912 8.328,9.584 7.5,9.584C6.672,9.584 6,8.912 6,8.084ZM7.5,7.584C7.224,7.584 7,7.808 7,8.084C7,8.36 7.224,8.584 7.5,8.584C7.776,8.584 8,8.36 8,8.084C8,7.808 7.776,7.584 7.5,7.584Z"
+      android:fillColor="#0173FF"
+      android:fillType="evenOdd"/>
+</vector>
