+++ app/src/main/res/drawable/ic_location.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="19dp"
+    android:height="19dp"
+    android:viewportWidth="19"
+    android:viewportHeight="19">
+  <path
+      android:pathData="M7.6,8.228C7.6,7.181 8.451,6.33 9.5,6.33C10.549,6.33 11.4,7.181 11.4,8.228C11.4,9.276 10.549,10.127 9.5,10.127C8.451,10.127 7.6,9.276 7.6,8.228Z"
+      android:fillColor="#9B9B9B"/>
+  <path
+      android:pathData="M1.267,8.228C1.267,3.684 4.954,0 9.5,0C14.046,0 17.733,3.684 17.733,8.228C17.733,11.616 16.038,14.299 14.223,16.113C13.313,17.022 12.36,17.727 11.543,18.208C11.135,18.449 10.754,18.638 10.423,18.769C10.11,18.892 9.781,18.988 9.5,18.988C9.219,18.988 8.89,18.892 8.577,18.769C8.246,18.638 7.864,18.449 7.456,18.208C6.64,17.727 5.687,17.022 4.777,16.113C2.962,14.299 1.267,11.616 1.267,8.228ZM9.5,5.063C7.752,5.063 6.333,6.481 6.333,8.228C6.333,9.976 7.752,11.393 9.5,11.393C11.248,11.393 12.667,9.976 12.667,8.228C12.667,6.481 11.248,5.063 9.5,5.063Z"
+      android:fillColor="#9B9B9B"
+      android:fillType="evenOdd"/>
+</vector>
