+++ app/src/main/res/drawable/ic_navigate.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="63dp"
+    android:height="63dp"
+    android:viewportWidth="63"
+    android:viewportHeight="63">
+  <path
+      android:pathData="M17.426,0L45.575,0A17.426,17.426 0,0 1,63 17.426L63,45.575A17.426,17.426 0,0 1,45.575 63L17.426,63A17.426,17.426 0,0 1,0 45.575L0,17.426A17.426,17.426 0,0 1,17.426 0z"
+      android:fillColor="#0173FF"/>
+  <path
+      android:pathData="M42.38,20.455C42.674,20.329 43.015,20.395 43.242,20.621C43.468,20.847 43.533,21.189 43.407,21.483L34.019,43.389C33.891,43.687 33.592,43.876 33.268,43.862C32.943,43.849 32.661,43.636 32.558,43.328L29.552,34.311L20.535,31.305C20.227,31.202 20.014,30.92 20.001,30.595C19.987,30.271 20.176,29.972 20.474,29.844L42.38,20.455Z"
+      android:fillColor="#ffffff"/>
+</vector>
