+++ app/src/main/res/drawable/ic_phone.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="17dp"
+    android:height="17dp"
+    android:viewportWidth="17"
+    android:viewportHeight="17">
+  <path
+      android:pathData="M2.833,0C1.269,0 0,1.269 0,2.833V5.1C0,11.672 5.328,17 11.9,17H14.167C15.731,17 17,15.731 17,14.167V12.6C17,11.957 16.636,11.368 16.06,11.08L13.322,9.711C12.397,9.248 11.276,9.712 10.949,10.694L10.612,11.706C10.483,12.091 10.089,12.32 9.691,12.24C7.203,11.743 5.257,9.797 4.76,7.309C4.68,6.911 4.909,6.517 5.294,6.388L6.55,5.969C7.393,5.689 7.877,4.806 7.662,3.944L6.998,1.288C6.808,0.531 6.129,0 5.348,0H2.833Z"
+      android:fillColor="#9B9B9B"/>
+</vector>
